+++ src/guards/is_assignment_pattern.ts
@@ -5,10 +5,14 @@ import { isLiteral } from './is_literal'
 type Node = TSESTree.Node
 type AssignmentPattern = TSESTree.AssignmentPattern
 
-export function isAssignmentPattern(node: Node, value?: string): node is AssignmentPattern {
+export function isAssignmentPattern(
+  node: Node,
+  value?: string
+): node is AssignmentPattern {
   return (
     node.type === AST_NODE_TYPES.AssignmentPattern &&
     (value === undefined ||
-      (node.right !== undefined && (isLiteral(node.right, value) || isIdentifier(node.right, value))))
+      (node.right !== undefined &&
+        (isLiteral(node.right, value) || isIdentifier(node.right, value))))
   )
 }