+++ src/guards/is_logical_expression.ts
@@ -3,6 +3,12 @@ import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/typescript-estree'
 type Node = TSESTree.Node
 type LogicalExpression = TSESTree.LogicalExpression
 
-export function isLogicalExpression(node: Node, operator?: string): node is LogicalExpression {
-  return node.type === AST_NODE_TYPES.LogicalExpression && (operator === undefined || node.operator === operator)
+export function isLogicalExpression(
+  node: Node,
+  operator?: string
+): node is LogicalExpression {
+  return (
+    node.type === AST_NODE_TYPES.LogicalExpression &&
+    (operator === undefined || node.operator === operator)
+  )
 }