+++ src/queries/find_first_of_type.ts
@@ -3,7 +3,10 @@ import { findFirst } from './find_first'
 
 type Node = TSESTree.Node
 
-export function findFirstOfType<T extends Node = Node>(root: Node, type: Node['type']): T | undefined {
+export function findFirstOfType<T extends Node = Node>(
+  root: Node,
+  type: Node['type']
+): T | undefined {
   const isOfType = (node: Node): node is T => node.type === type
   return findFirst(root, isOfType)
 }