+++ src/queries/find_raw_literal.ts
@@ -6,6 +6,7 @@ type Node = TSESTree.Node
 type Literal = TSESTree.Literal
 
 export function findRawLiteral(root: Node, raw: string): Literal | undefined {
-  const isLiteralValue = (node: Node): node is Literal => isLiteral(node) && node.raw === raw
+  const isLiteralValue = (node: Node): node is Literal =>
+    isLiteral(node) && node.raw === raw
   return findFirst(root, isLiteralValue) as Literal | undefined
 }